+++ tower-batch-control/tests/ed25519.rs
@@ -223,6 +223,30 @@ where
     Ok(())
 }
 
+async fn sign_and_verify_after_explicit_flush(
+    mut verifier: Batch<Verifier, Item>,
+    n: usize,
+) -> Result<(), BoxError> {
+    let mut results = FuturesOrdered::new();
+    for _ in 0..n {
+        let sk = SigningKey::new(thread_rng());
+        let vk_bytes = VerificationKeyBytes::from(&sk);
+        let msg = b"BatchVerifyTest";
+        let sig = sk.sign(&msg[..]);
+
+        verifier.ready().await?;
+        results.push_back(verifier.call((vk_bytes, sig, msg).into()));
+    }
+
+    verifier.flush()?;
+
+    while let Some(result) = results.next().await {
+        result?;
+    }
+
+    Ok(())
+}
+
 #[tokio::test(flavor = "multi_thread")]
 async fn batch_flushes_on_max_items_weight() -> Result<(), Report> {
     use tokio::time::timeout;
@@ -255,6 +279,25 @@ async fn batch_flushes_on_max_latency() -> Result<(), Report> {
         .await
         .map_err(|e| eyre!(e))?
         .map_err(|e| eyre!(e))?;
+
+    Ok(())
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn batch_flushes_on_explicit_flush() -> Result<(), Report> {
+    use tokio::time::timeout;
+    let _init_guard = zebra_test::init();
+
+    // Use a very high max_items and long max_latency. Without the explicit
+    // flush, this verification would wait for the latency timer.
+    let verifier = Batch::new(Verifier::default(), 100, 10, Duration::from_secs(1000));
+    timeout(
+        Duration::from_secs(1),
+        sign_and_verify_after_explicit_flush(verifier, 10),
+    )
+    .await
+    .map_err(|e| eyre!(e))?
+    .map_err(|e| eyre!(e))?;
 
     Ok(())
 }